Fixed relative path to colorscheme in .Xresources
Apparently include paths are to be relative to the .Xresources file, not
the symbolic link. So some changes had to be made to conform with
colors being in a separate package. Would probably be nice to create a
subfolder .colors for Xresources and a .colors for zsh at some point.